Todo app build update and fixes

+++ todo-app/src/app/api/[id]/route.ts
@@ -1,30 +1,53 @@
 import sqlite3 from 'sqlite3'
-import { open } from 'sqlite'
+import { Database, open } from 'sqlite'
+import { NextRequest, NextResponse } from 'next/server'
 
-let db = null
+let db: Database<sqlite3.Database, sqlite3.Statement> | null = null
 const filename: string = './src/database/todo.sqlite'
 
 export async function DELETE(
-	req: Request,
+	req: NextRequest,
 	{ params }: { params: { id?: number } }
-) {
+): Promise<Response> {
 	if (!db) {
 		db = await open({
 			filename,
 			driver: sqlite3.Database,
 		})
+	} else {
+		throw new Error('Database connection could not be established')
+	}
+
+	if (
+		typeof params.id !== 'number' ||
+		params.id === null ||
+		params.id === undefined
+	) {
+		return NextResponse.json(
+			{
+				error: { message: 'Invalid ID' },
+			},
+			{ status: 500 }
+		)
 	}
 
 	try {
 		await db.run('DELETE FROM todo WHERE id = ?', params.id)
+
+		return NextResponse.json(
+			{
+				message: 'Success',
+			},
+			{ status: 200 }
+		)
 	} catch (err) {
 		console.log(`Error: ${err}`)
-	}
 
-	return new Response(
-		JSON.stringify(
-			{ message: 'success' },
-			{ headers: { 'content-type': 'application/json' }, status: 200 }
+		return NextResponse.json(
+			{
+				error: { message: err },
+			},
+			{ status: 500 }
 		)
-	)
+	}
 }
